+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - Update method to IssuingPurchase resource
 ### Fixed
 - MerchantCategoryNumber type in IssuingPurchase resource, the API returns a string
+- BusinessAccountRequest resource
 
 ## [1.2.0] - 2026-07-03
 ### Fixed